fix(entries): use data table like in other pages and refactor
